requirements.txt
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
certifi==2018.10.15
channels==2.1.5
chardet==3.0.4
constantly==15.1.0
daphne==2.2.3
Django==2.1.2
django-filter==2.0.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
PyHamcrest==1.9.0
pyserial==3.4
pytz==2018.6
requests==2.20.0
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24
gunicorn==19.7.1
zope.interface==4.6.0
